// src/types/personalization.ts
export interface UserInterests {
  sports: {
    football: boolean;
    cricket: boolean;
    basketball: boolean;
    'american-football': boolean;
    tennis: boolean;
    formula1: boolean;
    rugby: boolean;
  };
  finance: {
    stocks: boolean;
    forex: boolean;
    crypto: boolean;
    commodities: boolean;
    realestate: boolean;
    startups: boolean;
  };
  technology: {
    fintech: boolean;
    ai: boolean;
    blockchain: boolean;
    cybersecurity: boolean;
    cloud: boolean;
    iot: boolean;
  };
  entertainment: {
    movies: boolean;
    music: boolean;
    tv: boolean;
    books: boolean;
    gaming: boolean;
  };
  lifestyle: {
    food: boolean;
    travel: boolean;
    health: boolean;
    fitness: boolean;
    culture: boolean;
  };
  regions: {
    local: boolean;
    caribbean: boolean;
    global: boolean;
    northAmerica: boolean;
    europe: boolean;
    asia: boolean;
  };
}